Add files via upload
This is the progress you asked us to upload. I have been working on my camera for the last week and my promoter - Mr Szklanny - asked me to take some pictures of what is the quality if the camera and how the photos could look like. So, this is it :)